Fix a bug with GitHub pagination
When the GitHub account had not enough repositories to trigger
pagination, the 'Link' header would be absent and the
'parse-link-header' module would return 'null' instead of an expected
(even empty) object.

By forcing the argument to be a single space when the header is absent,
the module will return an empty object (so we're not effectively
accessing 'next' on 'null'). It's a single space because an empty string
would result in 'null' too.